Bump max packet size and core pinning.
Previously, all rx queues and tx queues would be pinned to the same CPU.
But it makes more sense to pin the rx queues to separate cores so the
kernel can distribute the work over multiple cores. Since we don't use
tx queues, we can just pin them to the same place.